Opprett tilganger service
Intensjonen er at den kan hente tilganger, passe på
at et feilende kall ikke påvirker andre, og returnere
dem i en passende datastruktur, som inneholder alle tilganger
som ble funnet, og feilene.
